pre pull

+++ protected/views/user/user_files.php
@@ -6,30 +6,30 @@
 
 <div class="mydropbox">
   <div class="test">
-
-    <div class="dop_menu">
-      <div class="files_upload" onclick="doorDownloadFile(event)">
-        <span>Загрузить</span>
+    <?php if ($mu_path): ?>
+      <div class="dop_menu">
+        <div class="files_upload" onclick="doorDownloadFile(event)">
+          <span>Загрузить</span>
+        </div>
+        <div class="files_newfolder" onclick="changeFolder(0, event)" title="Создать папку">
+          <span>Новая папка</span>
+        </div>
       </div>
-      <div class="files_newfolder" onclick="changeFolder(0, event)" title="Создать папку">
-        <span>Новая папка</span>
-      </div>
-    </div>
-
-    <ul class="ul_files_actions">
-      <li class="files_dowland">
-        <div></div>
-        <span>Скачать</span>
-      </li>
-      <li class="files_delete" onclick="deleteFolder(event)">
-        <div></div>
-        <span>Удалить</span>
-      </li>
-      <li class="files_rename" onclick="editLineFolder(event)" >
-        <div></div>
-        <span >Изменить</span>
-      </li>
-    </ul>
+      <ul class="ul_files_actions">
+        <!--        <li class="files_dowland">
+                  <div></div>
+                  <span>Скачать</span>
+                </li>-->
+        <li class="files_delete" onclick="deleteFolder(event)">
+          <div></div>
+          <span>Удалить</span>
+        </li>
+        <li class="files_rename" onclick="editLineFolder(event)" >
+          <div></div>
+          <span >Изменить</span>
+        </li>
+      </ul>
+    <?php endif; ?>
   </div>
   <div class="anchor"></div>
 
@@ -50,10 +50,12 @@
           <div class="block_panel"></div>
           <span>Изменено</span>
         </div>
-        <div class="td_t">
-          <div class="block_panel"></div>
-          <span>Область видимости</span>
-        </div>
+        <?php if ($mu_path): ?>
+          <div class="td_t">
+            <div class="block_panel"></div>
+            <span>Область видимости</span>
+          </div>
+        <?php endif; ?>
       </div>
     </div>
     <div class="table_body_t">
@@ -62,7 +64,8 @@
         echo $this->renderPartial('_folder', array(
             'folder' => $folder,
             'private_status' => $private_status,
-            'new' => $new
+            'new' => $new,
+            'mu_path' => $mu_path
                 ), true);
         ?>
       <?php endforeach; ?>
